+++ justify.go
@@ -85,24 +85,24 @@ func parseJustifyArgs(s *State, argsStr []rune, funcName string) (*JustifyArgs, 
 		return nil, err
 	}
 	if len(args) < 2 {
-		s.errorOffset += len(argsStr) + 1
+		s.errorOffset += int64(len(argsStr) + 1)
 		return nil, s.errorArg("%s: at least 2 arguments are required", funcName)
 	}
 	pattern := []rune(args[0])
 	width, err := strconv.Atoi(strings.TrimSpace(string(args[1])))
 	if err != nil {
-		s.errorOffset += len(args[0]) + len(args[1]) + 1
+		s.errorOffset += int64(len(args[0]) + len(args[1]) + 1)
 		return nil, s.errorValue("invalid width %s", string(args[1]))
 	}
 	if width < 1 {
-		s.errorOffset += len(args[0]) + len(args[1]) + 1
+		s.errorOffset += int64(len(args[0]) + len(args[1]) + 1)
 		return nil, s.errorValue("invalid width %s", string(args[1]))
 	}
 	fillChar := ' '
 	if len(args) > 2 {
 		fillCharA := []rune(args[2])
 		if len(fillCharA) != 1 {
-			s.errorOffset += len(args[0]) + len(args[1]) + len(args[2]) + 2
+			s.errorOffset += int64(len(args[0]) + len(args[1]) + len(args[2]) + 2)
 			return nil, s.errorValue("invalid fillChar=%#v, must have length 1", string(args[2]))
 		}
 		fillChar = fillCharA[0]
