+++ lex.go
@@ -117,12 +117,12 @@ func makeLexUnicode(parentLex LexType, symbol rune, width int, toBuff bool) LexT
 			buff = append(buff, s.pattern[s.patternPos])
 		}
 		if len(buff) != width {
-			s.errorOffset -= len(buff) - 1
+			s.errorOffset -= int64(len(buff) - 1)
 			return nil, s.errorSyntax("invalid escape sequence")
 		}
 		char, _, _, err := strconv.UnquoteChar(string(buff), '"')
 		if err != nil {
-			s.errorOffset -= width - 1
+			s.errorOffset -= int64(width - 1)
 			return nil, s.errorSyntax("invalid escape sequence")
 		}
 		if toBuff {
